python/requirements.txt
prometheus-client==0.8.0
fluent-logger==0.9.6
